pyproject.toml
[tool.poetry]
name = "hephaestus"
version = "0.1.0"
description = "an agent that can create and improve upon tools and components"
authors = ["Mengyang (John) Cha <15233508+solarapparition@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.343"
colorama = "^0.4.6"
openai = "^1.3.7"
ruamel-yaml = "^0.18.5"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.2"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
